services/agent-orchestrator/requirements.txt
functions-framework==3.*
google-cloud-aiplatform>=1.38.0
google-cloud-firestore>=2.14.0
google-cloud-pubsub>=2.18.0
google-cloud-bigquery>=3.11.0
google-cloud-storage>=2.10.0
pydantic>=2.0.0
python-dotenv>=1.0.0
tenacity>=8.2.0
